+++ src/FsCodec/StreamName.fs
@@ -87,6 +87,13 @@ module StreamName =
     (* Splitting: functions/Active patterns for (i.e. generated via `parse`, `create` or `compose`) well-formed Stream Names
        Will throw if presented with malformed strings [generated via alternate means] *)
 
+    /// Extracts the category portion of the StreamName
+    let category (x : StreamName) =
+        let raw = toString x
+        raw.Substring(0, raw.IndexOf '-')
+    /// Extracts the category portion of a StreamName
+    let (|Category|) = category
+
     /// <summary>Splits a well-formed Stream Name of the form <c>{category}-{streamId}</c> into its two elements.<br/>
     /// Throws <c>InvalidArgumentException</c> if it does not adhere to the well known format (i.e. if it was not produced by `parse`).</summary>
     /// <remarks>Inverse of <c>create</c></remarks>